requirements.txt
Django==4.2.7
Pillow==10.1.0
google-generativeai==0.3.2
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
plotly==5.17.0
pandas==2.1.4
numpy==1.24.3
pytz==2023.3
psycopg2-binary==2.9.9